+++ internal/texture/constant.go
@@ -1,6 +1,6 @@
 package texture
 
-import "github.com/flynn-nrg/izpi/internal/vec3"
+import "github.com/flynn-nrg/go-vfx/math32/vec3"
 
 // Ensure interface compliance.
 var _ Texture = (*Constant)(nil)
@@ -17,6 +17,6 @@ func NewConstant(color vec3.Vec3Impl) *Constant {
 	}
 }
 
-func (c *Constant) Value(_ float64, _ float64, _ vec3.Vec3Impl) vec3.Vec3Impl {
+func (c *Constant) Value(_ float32, _ float32, _ vec3.Vec3Impl) vec3.Vec3Impl {
 	return c.color
 }
